[Refactor](schema-form): Improve documentation and examples for type guards (isSchemaNode, isBooleanNode, isNumberNode, isObjectNode, isStringNode, isVirtualNode, isArrayNode, isBranchNode, isTerminalNode, isNullNode) to enhance usability and clarity in schema node validation
