+++ packages/canard/schema-form/src/core/nodes/ArrayNode/index.ts
@@ -1,1 +1,2 @@
 export { ArrayNode } from './ArrayNode';
+export { isArrayNode } from './filter';
