+++ packages/canard/schema-form/src/core/nodes/BooleanNode/index.ts
@@ -1,1 +1,2 @@
 export { BooleanNode } from './BooleanNode';
+export { isBooleanNode } from './filter';
