+++ packages/canard/schema-form/src/core/nodes/NullNode/index.ts
@@ -1,1 +1,2 @@
 export { NullNode } from './NullNode';
+export { isNullNode } from './filter';
