+++ packages/canard/schema-form/src/core/nodes/NumberNode/index.ts
@@ -1,1 +1,2 @@
 export { NumberNode } from './NumberNode';
+export { isNumberNode } from './filter';
