+++ packages/canard/schema-form/src/core/nodes/ObjectNode/index.ts
@@ -1,1 +1,2 @@
 export { ObjectNode } from './ObjectNode';
+export { isObjectNode } from './filter';
