+++ packages/canard/schema-form/src/core/nodes/StringNode/index.ts
@@ -1,1 +1,2 @@
 export { StringNode } from './StringNode';
+export { isStringNode } from './filter';
