+++ packages/canard/schema-form/src/core/nodes/VirtualNode/index.ts
@@ -1,1 +1,2 @@
 export { VirtualNode } from './VirtualNode';
+export { isVirtualNode } from './filter';
