+++ packages/canard/schema-form/src/core/nodes/index.ts
@@ -9,15 +9,12 @@ export {
 } from './type';
 export { createSchemaNodeFactory } from './schemaNodeFactory';
 
-export {
-  isSchemaNode,
-  isBooleanNode,
-  isNumberNode,
-  isObjectNode,
-  isStringNode,
-  isVirtualNode,
-  isArrayNode,
-  isBranchNode,
-  isTerminalNode,
-  isNullNode,
-} from './filter';
+export { isArrayNode } from './ArrayNode';
+export { isBooleanNode } from './BooleanNode';
+export { isNullNode } from './NullNode';
+export { isNumberNode } from './NumberNode';
+export { isObjectNode } from './ObjectNode';
+export { isStringNode } from './StringNode';
+export { isVirtualNode } from './VirtualNode';
+
+export { isSchemaNode, isBranchNode, isTerminalNode } from './filter';
